requirements.txt
django-transmeta==0.6.2
